Completely revamped the point selection process
The point selection process is not based on the surf data directly,
but rather the axes box outer points are used.

This version does not work with plots that have only one of the
viewing angles differ from 0. But a lot has changed, like the stuff
regarding the boxes. Also the solution to find the appropriate points
is too cute to not save it somewhere, although it will be replaced in
the next commit.